+++ macros/src/main/scala/goggles/macros/interpret/Contextual.scala
@@ -1,0 +1,9 @@
+package goggles.macros.interpret
+
+import scala.reflect.macros.whitebox
+
+trait Contextual {
+  val c: whitebox.Context
+  
+  type Interpret[A] = Parse[c.Type, c.Expr[Any], A]
+}
