+++ macros/src/main/scala/goggles/macros/interpret/DslModeContext.scala
@@ -1,0 +1,5 @@
+package goggles.macros.interpret
+
+trait DslModeContext {
+  def mode: DslMode
+}
