+++ macros/src/main/scala/goggles/macros/interpret/MacroResult.scala
@@ -2,8 +2,8 @@ package goggles.macros.interpret
 
 import goggles.macros.errors.GogglesError
 
-private[goggles] case class MacroResult[+Type, +Tree](
-  errorOrResult: Either[GogglesError[Type], Tree],
+private[goggles] case class MacroResult[+Type, +A](
+  errorOrResult: Either[GogglesError[Type], A],
   infos: List[OpticInfo[Type]],
   lastSegmentOffset: Int)
