+++ macros/src/main/scala/goggles/macros/interpret/StringContextInterpreter.scala
@@ -1,0 +1,16 @@
+package goggles.macros.interpret
+
+
+trait StringContextInterpreter {
+  self: Contextual =>
+  
+  def contextStringParts: List[String] = {
+    import c.universe._
+  
+    c.prefix.tree match {
+      case Apply(f, List(Apply(g, rawParts))) => rawParts.map {
+        case Literal(Constant(str: String)) => str
+      }
+    }
+  }
+}
